Lay framework for rejecting invalid note hits
This deals with rejecting any taps that are not correct for the note. That may be things such as a hold for an expected tap, sliding across lanes, not holding long enough, and not tapping in proper lane.

Still waiting on note spawning, to be able to get the current note.